// src/edid-vendor-types.ts
export type EdidVendorBrand =
  | 'Acer'
  | 'Apple'
  | 'Asus'
  | 'BenQ'
  | 'Dell'
  | 'Hisense'
  | 'HP'
  | 'IBM'
  | 'Lenovo'
  | 'LG'
  | 'MSI'
  | 'Panasonic'
  | 'Philips' // aka TP Vision
  | 'Samsung'
  | 'Sharp'
  | 'Sony'
  | 'TCL'
  | 'Toshiba'
  | 'Unknown'
  | 'ViewSonic'
  | 'Vizio';
// ...
export type EdidVendorId = keyof EdidVendorIdMap;

export type EdidVendorName = ValueOf<{
  [K in EdidVendorId]: EdidVendorIdMap[K] extends string
    ? EdidVendorIdMap[K]
    : EdidVendorIdMap[K][number];
}>;